Datastore Extraction: StVarName refactor, small non-functional change in furtherance of datastore extraction (#2795)
